backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
mysqlclient>=2.2,<3.0
gunicorn>=22.0,<23.0
Pillow>=10.0,<11.0
whitenoise>=6.5,<7.0
